pyproject.toml
[tool.poetry]
name = "arcade_html2pdf"
version = "0.0.1"
description = "Tools to interact with html2pdf api, now Arcade can download webpages as PDFs!"
authors = ["santiagoSSAA <santiago.sanchez.ssaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "0.1.*"
httpx = "^0.28.1"
python-dotenv="^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"
pytest-cov = "^4.0.0"
pytest-asyncio="^0.24.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["arcade_html2pdf/**/*.py"]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true
